+++ src/parser/block/paragraph/mention/address.ts
@@ -2,8 +2,8 @@ import { ParagraphParser } from '../../../block';
 import { union, some, line, rewrite, focus, match, convert, trim, fmap } from '../../../../combinator';
 import { link } from '../../../inline';
 import { unescsource } from '../../../source/unescapable';
-import { html, define } from 'typed-dom';
 import { memoize } from '../../../util';
+import { define } from 'typed-dom';
 
 export const address: ParagraphParser.MentionParser.AddressParser = line(
   fmap(
@@ -18,7 +18,11 @@ export const address: ParagraphParser.MentionParser.AddressParser = line(
           ]),
           convert(source => `[${prefix}]{ ${source} }`, union([link])))))),
     ([el]) => [
-      html('span', { class: 'address', 'data-level': `${el.textContent!.length}` }, [
-        define(el, { href: null }, `${el.textContent}${el.getAttribute('href')}`)
-      ])
+      define(el,
+        {
+          class: 'address',
+          'data-level': `${el.textContent!.length}`,
+          href: null
+        },
+        `${el.textContent}${el.getAttribute('href')}`)
     ]));